Split inductor utils helpers into focused modules
