+++ torch/_inductor/codegen/cutlass/utils.py
@@ -6,6 +6,7 @@ import os
 import shutil
 import sys
 import time
+import warnings
 from dataclasses import dataclass
 from typing import Any
 from typing_extensions import TypeIs
@@ -14,7 +15,11 @@ import sympy
 
 import torch
 from torch._inductor.runtime.runtime_utils import dynamo_timed
-from torch._inductor.utils import clear_on_fresh_cache
+from torch._inductor.utils import (
+    _use_autotune_backend,
+    _use_template_for_gpu,
+    clear_on_fresh_cache,
+)
 from torch.utils._ordered_set import OrderedSet
 
 from ... import config
@@ -32,6 +37,50 @@ ACCUMULATOR_DTYPES: OrderedSet[torch.dtype] = OrderedSet([torch.float, torch.int
 XW_DTYPES: OrderedSet[torch.dtype] = OrderedSet(
     [torch.half, torch.bfloat16, torch.float8_e4m3fn, torch.int8, torch.float8_e5m2]
 )
+
+
+def use_cutlass_template(layout: Layout, m: int, n: int, k: int) -> bool:
+    # TODO: Enable CUTLASS in non-AOT cpp_wrapper mode. The CUTLASS
+    # codegen (CUDATemplateKernel.call_kernel) already has cpp_wrapper-aware
+    # arg handling, but the other half is missing: the non-triton branch of
+    # CppWrapperGpu._generate_kernel_call_helper unconditionally emits
+    # `kernels.<name>(...)`, and that AOTInductorModelKernels struct only
+    # exists in AOT mode. Fixing this requires adding dlopen/dlsym loading
+    # for the compiled CUTLASS .so, similar to how the triton branch uses
+    # static CUfunction + loadKernel for non-AOT mode.
+    if V.graph.cpp_wrapper and not V.graph.aot_mode:
+        warnings.warn(
+            "CUTLASS backend is not supported with non-AOT cpp_wrapper mode. "
+            "Skipping CUTLASS backend.",
+        )
+        return False
+
+    gemm_size = V.graph.sizevars.optimization_hint(m * n * k, fallback=-1)
+    if gemm_size <= 0 or gemm_size < config.cutlass.cutlass_backend_min_gemm_size:
+        return False
+
+    # Do not use cutlass template on ROCm
+    if torch.version.hip:
+        return False
+
+    # output dtype
+    # FP32 not supported: https://github.com/pytorch/pytorch/issues/145952
+    layout_dtypes = [torch.float16, torch.bfloat16, torch.int32]
+    res = (
+        _use_template_for_gpu(layout, layout_dtypes)
+        and (config.max_autotune or config.max_autotune_gemm)
+        and _use_autotune_backend("CUTLASS")
+    )
+
+    if res and not try_import_cutlass():
+        log.warning(
+            "Failed to import CUTLASS lib. Please check whether "
+            "_inductor.config.cutlass.cutlass_dir %s is set correctly. "
+            "Skipping CUTLASS backend for now.",
+            config.cutlass.cutlass_dir,
+        )
+        return False
+    return res
 
 
 @atexit.register
