+++ torch/_inductor/codegen/rocm/ck_template.py
@@ -1,12 +1,84 @@
+import logging
 from typing import Any
 from typing_extensions import override
 
 import torch
 from torch._inductor.codegen.rocm.rocm_template import ROCmTemplate
-from torch._inductor.ir import IRNode
-from torch._inductor.utils import IndentedBuffer
+from torch._inductor.code_formatting import IndentedBuffer
+from torch._inductor.ir import IRNode, Layout
+from torch._inductor.utils import (
+    _rocm_native_device_arch_name,
+    _use_autotune_backend,
+    _use_conv_autotune_backend,
+    try_import_ck_lib,
+)
 
 from .rocm_template import ArgInfo
+
+log = logging.getLogger(__name__)
+
+
+def use_ck_template(layout: Layout) -> bool:
+    from ... import config
+
+    # config knobs check 1
+    if not (config.max_autotune or config.max_autotune_gemm):
+        return False
+    # platform check
+    if not torch.version.hip:
+        return False
+    # tensors must be on GPU
+    if layout.device.type != "cuda":
+        return False
+    # hardware check
+    # if config arch list is not specified, get the native arch from the device properties
+    native_arch = _rocm_native_device_arch_name(layout.device)
+    requested_archs = {k.split(":")[0]: k for k in config.rocm.arch} or {
+        native_arch.split(":")[0]: native_arch
+    }
+    requested_supported_archs = [
+        requested_archs[k]
+        for k in requested_archs.keys() & config.rocm.ck_supported_arch
+    ]
+    if not requested_supported_archs:
+        return False
+    # supported input dtypes
+    if layout.dtype not in [torch.float16, torch.bfloat16, torch.float32]:
+        return False
+
+    ck_package_dirname, _, _, _ = try_import_ck_lib()
+
+    if not ck_package_dirname:
+        log.warning("Please pip install Composable Kernel package")
+        return False
+
+    config.rocm.ck_dir = ck_package_dirname
+
+    return True
+
+
+def use_ck_gemm_template(layout: Layout, m: int, n: int, k: int) -> bool:
+    from ...virtualized import V
+
+    return (
+        _use_autotune_backend("CK")
+        and use_ck_template(layout)
+        and V.graph.sizevars.optimization_hint(m * n * k, fallback=-1) > 0
+    )
+
+
+def use_ck_tile_gemm_template(layout: Layout, m: int, n: int, k: int) -> bool:
+    from ...virtualized import V
+
+    return (
+        _use_autotune_backend("CKTILE")
+        and use_ck_template(layout)
+        and V.graph.sizevars.optimization_hint(m * n * k, fallback=-1) > 0
+    )
+
+
+def use_ck_conv_template(layout: Layout) -> bool:
+    return _use_conv_autotune_backend("CK") and use_ck_template(layout)
 
 
 class CKTemplate(ROCmTemplate):
